feat: standardize how dates are handled.
updateDttm is now stored in the database as an unsigned int and assigned by the service layer.

event.beginDttm and event.endDttm are still stored in the database as DateTime values, but are now assumed to be stored as UTC. The API expects these to be in ISO 8601 UTC format, milliseconds optional, and the service layer strips the T and Z delimiter to be able to pass a "SQL date."

Since the API has a different date format than the storage, there are separate model JSON schemas for each. The API schemas drive the swagger doc and don't require updateDttm or updateUser. The db schemas are used to actually validate the data before saving.